requirements.txt
streamlit
Pillow
numpy
requests
python-dotenv
google-generativeai
chromadb
datasets
